third_party/WebKit/LayoutTests/fast/ime/html-inputmethodcontext-expected.txt

This tests inputMethodContext attribute which is an additional interface for HTMLElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new InputMethodContext() threw exception TypeError: Illegal constructor.
PASS [object InputMethodContext] is non-null.
PASS typeof(html.inputMethodContext) is "object"
PASS context instanceof InputMethodContext is true
PASS context2 instanceof InputMethodContext is true
PASS context === context2 is true
PASS inputContext is not input2Context
PASS successfullyParsed is true

TEST COMPLETE
